Add a `{'fun', fun()}` retention spec
This can be used flexibly to evaluate retention depending on the name or
contents of index files. You pass in a function which returns a tuple
with the index files split into two lists: to delete and to keep.
This could be used as a way to truncate everything up to an offset or to
guarantee that an offset (for example an uncommitted one) won't be
truncated. Since these files are sorted, some retention functions could
operate just on the names (deriving the offset of the segment with
`erlang:binary_to_integer/1`).