osiris_writer: Track diff between writer and replica offsets
This introduces a metric calculated at every batch which records the
difference between the last known offset of the writer and its replicas.
This can be used to observe a backlog of replication diminish, or to
diagnose situations when replication is being starved out (network-wise)
by high-throughput publishing. The replication diff should usually be
low but, for streams seeing traffic, non-zero.